make gray_image return correct dtype
camera.gray_image now preserves the data type of the raw snapshot when
converting from colour to grayscale

+++ nplab/instrument/camera/__init__.py
@@ -182,7 +182,7 @@ class Camera(Instrument, HasTraits):
         except:
             try:
                 assert frame.shape[2]==3
-                return np.mean(frame,axis=2)
+                return np.mean(frame, axis=2, dtype=frame.dtype)
             except:
                 return None
